Adds more user templates
