--- broker_tom/alerts/urls.py
@@ -1,13 +1,0 @@
-# !/usr/bin/env python3
-# -*- coding: UTF-8 -*-
-
-"""django URL Configuration. Lists routes from URLs to views"""
-
-from django.urls import path
-
-from .views import AlertSummaryView, AlertsView
-
-urlpatterns = [
-    path('', AlertsView.as_view(), name='alerts'),
-    path('<int:pk>', AlertSummaryView.as_view(), name='alert-summary'),
-]
