+++ broker_tom/broker_tom/urls.py
@@ -9,13 +9,15 @@ from django.contrib import admin
 from django.contrib.auth.views import LoginView, LogoutView
 from django.urls import include, path
 
-from .views import IndexView
+from .views import AlertSummaryView, AlertsView, IndexView
 
 # URLs for custom views
 urlpatterns = [
     path('', IndexView.as_view(), name='home'),
-    path('alerts/', include('alerts.urls')),
+    path('alerts/', AlertsView.as_view(), name='alerts'),
+    path('alerts/<int:pk>', AlertSummaryView.as_view(), name='alert-summary'),
     path('gettingstarted/', include('getting_started.urls')),
+    path('user/', include('user_subscriptions.urls')),
 ]
 
 # Built in Django URL patterns
