+++ broker_tom/templates/broker_tom/alert_summary.html
@@ -1,5 +1,4 @@
 {% extends 'base.html' %}
-{% load bootstrap4 %}
 
 {% block title %}Alert Summary{% endblock %}
 
@@ -24,7 +23,9 @@
 
   <br>
 
-  <h3>Alert Data</h3>
+  <div class="section-header">
+    <h3>Alert Data</h3>
+  </div>>
   <p>Alert data is presented as provided by the parent survey as is not
   modified in any way.
   </p>
@@ -43,7 +44,9 @@
 
   <br>
 
-  <h3>Value Added Data Products</h3>
+  <div class="section-header">
+    <h3>Value Added Data Products</h3>
+  </div>
   <p>Value added data products may not be available for alerts that are still
   undergoing processing. Furthermore, some data products are determined
   conditionally and are only available for certain alerts. See
