+++ broker_tom/templates/broker_tom/alerts.html
@@ -1,5 +1,4 @@
 {% extends 'base.html' %}
-{% load bootstrap4 %}
 
 {% block title %}Alerts{% endblock %}
