+++ broker_tom/templates/getting_started/introduction.html
@@ -1,9 +1,7 @@
-{% extends 'getting_started/base.html' %}
-{% load bootstrap4 %}
+{% extends 'getting_started/side_bar_menu_base.html' %}
+
 {% block title %}Getting Started{% endblock %}
 {% block main-column %}
-<h3>Using PGB Data Products</h3>
-
-
-This page will demonstrate how to access and interface with the data in the cloud.
+  <h3>Using PGB Data Products</h3>
+  This page will demonstrate how to access and interface with the data in the cloud.
 {% endblock %}
