+++ broker_tom/templates/getting_started/technical_resources.html
@@ -1,4 +1,4 @@
-{% extends 'getting_started/base.html' %}
+{% extends 'getting_started/side_bar_menu_base.html' %}
 
 {% block title %}Technical Resources{% endblock %}
