+++ broker_tom/templates/navbar_content.html
@@ -1,4 +1,3 @@
-{% load bootstrap4 %}
 <li class="nav-item {% if request.resolver_match.url_name == 'home' %}active{% endif %}">
   <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
 </li>
@@ -14,6 +13,6 @@
 
 {% if user.is_authenticated %}
   <li class="nav-item {% if request.resolver_match.url_name == 'user' %}active{% endif %}">
-    <a class="nav-link" href="https://pitt-broker.readthedocs.io/en/latest/">My Profile</a>
+    <a class="nav-link" href="{% url 'user' %}">My Profile</a>
   </li>
 {% endif %}
