+++ broker_tom/templates/registration/change_user_password.html
@@ -1,0 +1,17 @@
+{% extends 'tom_common/base.html' %}
+
+{% block title %}Change password{% endblock %}
+
+{% block content %}
+  <div style="text-align:center">
+    <form action="{% url 'admin-user-change-password' view.kwargs.pk %}" method="POST">
+      {% csrf_token %}
+      {% bootstrap_form form %}
+      {% buttons %}
+        <button type="submit" class="btn btn-primary">
+          Change Password
+        </button>
+      {% endbuttons %}
+    </form>
+  </div>
+{% endblock %}
