+++ broker_tom/templates/registration/logged_out.html
@@ -2,5 +2,7 @@
 
 {% block content %}
   <meta http-equiv="Refresh" content="3; url={% url 'home' %}" />
-  <p>You have been successfully logged out. You will be automatically redirected in 3 seconds...</p>
+  <div style="text-align:center">
+    <p>You have been successfully logged out. You will be automatically redirected in 3 seconds...</p>
+  </div>
 {% endblock %}
