+++ broker_tom/templates/registration/login.html
@@ -1,37 +1,31 @@
 {% extends "base.html" %}
 
 {% block content %}
-
-  {% if form.errors %}
-    <p>The provided username and password did not match. Please try again.</p>
+  {% if user.is_authenticated %}
+    <meta http-equiv="Refresh" content="0; url={% url 'home' %}" />
   {% endif %}
 
-  {% if next %}
-    {% if user.is_authenticated %}
-      <p>Your account doesn't have access to this page. To proceed,
-      please login with an account that has access.</p>
-    {% else %}
-      <p>Please login to see this page.</p>
+  <div style="text-align:center">
+    <h2> Login </h2>
+    {% if form.errors %}
+      <p>The provided username and password did not match. Please try again.</p>
     {% endif %}
-  {% endif %}
 
-  <form method="post" action="{% url 'login' %}">
-    {% csrf_token %}
-    <table>
-      <tr>
-        <td>{{ form.username.label_tag }}</td>
-        <td>{{ form.username }}</td>
-      </tr>
-      <tr>
-        <td>{{ form.password.label_tag }}</td>
-        <td>{{ form.password }}</td>
-      </tr>
-    </table>
-    <input type="submit" value="login" />
-    <input type="hidden" name="next" value="{{ next }}" />
-  </form>
-
-  {# Assumes you setup the password_reset view in your URLconf #}
-  <p><a href="{% url 'password_reset' %}">Lost password?</a></p>
-
+    <form method="post" action="{% url 'login' %}" class="form-group">
+      {% csrf_token %}
+      <table>
+        <tr>
+          <td>{{ form.username.label_tag }}</td>
+          <td>{{ form.username }}</td>
+        </tr>
+        <tr>
+          <td>{{ form.password.label_tag }}</td>
+          <td>{{ form.password }}</td>
+        </tr>
+      </table>
+      <button type="submit" class="btn btn-primary">Login</button>
+      <input type="button" class="btn btn-secondary" onclick="window.location.href = '{% url 'password_reset' %}';" value="Forgot your password?"/>
+      <input type="hidden" name="next" value="{{ next }}" />
+    </form>
+  </div>
 {% endblock %}
