+++ broker_tom/templates/registration/password_reset_complete.html
@@ -1,6 +1,8 @@
 {% extends "base.html" %}
 
 {% block content %}
-  <h2>Congratulations, your has been changed successfully!</h2>
-  <p><a href="{% url 'login' %}">log in again?</a></p>
+  <div style="text-align:center">
+    <h2>Congratulations, your has been changed successfully!</h2>
+    <p><a href="{% url 'login' %}">log in again?</a></p>
+  </div>
 {% endblock %}
