+++ broker_tom/templates/registration/password_reset_done.html
@@ -1,6 +1,8 @@
 {% extends "base.html" %}
 
 {% block content %}
+  <div style="text-align:center">
     <p>Instructions for resetting your password have been sent to your email.</p>
     <p>If you do not receive an email in the next few minutes, please check your spam folder.</p>
+  </div>
 {% endblock %}
