+++ broker_tom/templates/registration/password_reset_form.html
@@ -1,12 +1,15 @@
 {% extends "base.html" %}
 
 {% block content %}
-  <form action="" method="post">
-  {% csrf_token %}
-  {% if form.email.errors %}
-    {{ form.email.errors }}
-  {% endif %}
-      <p>{{ form.email }}</p>
-    <input type="submit" class="btn btn-default btn-lg" value="Reset password">
-  </form>
+  <div style="text-align:center">
+    <p>Please enter your email and we will send you a link to reset your password.</p>
+    <form action="" method="post">
+    {% csrf_token %}
+    {% if form.email.errors %}
+      {{ form.email.errors }}
+    {% endif %}
+      <p>Email: {{ form.email }}</p>
+      <button type="submit" class="btn btn-primary">Reset password</button>
+    </form>
+  </div>
 {% endblock %}
