+++ broker_tom/templates/user_subscriptions/edit_subscriptions.html
@@ -1,0 +1,9 @@
+{% extends 'base.html' %}
+{% load static %}
+
+{% block title %}My Subscriptions{% endblock %}
+
+{% block content %}
+  <h2>My subscriptions</h2>
+
+{% endblock %}
