+++ broker_tom/templates/user_subscriptions/my_profile.html
@@ -1,0 +1,58 @@
+{% extends 'base.html' %}
+{% load static %}
+
+{% block title %}My Profile{% endblock %}
+
+{% block content %}
+  <h2>{{user.first_name}} {{user.last_name}} David Tennant</h2>
+  <ul class="list-unstyled">
+    <li> <b>Email:</b> {% if user.email %}{{user.email}}{% else %}None{% endif %} </li>
+    <li> <b>Affiliate University:</b> {% if user.university %}{{user.university}}{% else %}None{% endif %} </li>
+  </ul>
+
+  <br>
+
+  <div class="section-header">
+    <h3>My Topic Subscriptions</h3>
+    <span><a href="{% url 'edit-subscriptions' %}">(Edit)</a></span>
+  </div>
+
+  <br>
+  <div class="section-header">
+    <h3> Submitted Observations </h3>
+  </div>
+  <table class="table table-striped">
+    <tr>
+      <th>Timestamp</th>
+      <th>Observatory</th>
+      <th>Status</th>
+    </tr>
+    {% for observation in submitted_observations %}
+    <tr>
+      <td></td>
+      <td></td>
+      <td></td>
+    </tr>
+    {% endfor %}
+  </table>
+
+  <br>
+
+  <div class="section-header">
+    <h3> Recent Subscribed Alerts </h3>
+  </div>
+  <table class="table table-striped">
+    <tr>
+      <th>Timestamp</th>
+      <th>Topic</th>
+      <th>Alert</th>
+    </tr>
+    {% for time, message in subscribed_alerts %}
+    <tr>
+      <td><a href="{% url 'alerts' %}{{time}}">{{time}}</a></td>
+      <td>Some topic</td>
+      <td>message</td>
+    </tr>
+    {% endfor %}
+  </table>
+{% endblock %}
